Eliminated a lot of crud by making non-generic versions rely on generic ones
